backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
PyPDF2==3.0.1
dashscope==1.17.0
python-dotenv==1.0.0
jieba==0.42.1
scikit-learn==1.3.0
numpy==1.24.3
sentence-transformers
faiss-cpu>=1.7.0
torch>=1.13.0
